// src/generated/wotlk/api.ts
export interface Party {
    /**
     * @generated from protobuf field: repeated proto.Player players = 1
     */
    players: Player[];
    /**
     * @generated from protobuf field: proto.PartyBuffs buffs = 2
     */
    buffs?: PartyBuffs;
}
/**
 * @generated from protobuf message proto.Raid
 */
export interface Raid {
    /**
     * @generated from protobuf field: repeated proto.Party parties = 1
     */
    parties: Party[];
    /**
     * The number of parties in the raid who participate in the sim, e.g.
     * 10-man vs 25-man raids. The remaining parties are unused but can hold
     * bench players.
     *
     * If this value is 0, all parties are included.
     *
     * @generated from protobuf field: int32 num_active_parties = 2
     */
    numActiveParties: number;
    /**
     * @generated from protobuf field: proto.RaidBuffs buffs = 7
     */
    buffs?: RaidBuffs;
    /**
     * Extra debuffs provided by buff bots in this raid.
     * This is also used for debuffs in the individual sims.
     *
     * @generated from protobuf field: proto.Debuffs debuffs = 5
     */
    debuffs?: Debuffs;
    /**
     * Players who will be tanking mobs.
     *
     * @generated from protobuf field: repeated proto.UnitReference tanks = 4
     */
    tanks: UnitReference[];
    /**
     * Staggers Stormstrike casts across Enhance Shaman to maximize charge usage.
     *
     * @generated from protobuf field: bool stagger_stormstrikes = 3
     */
    staggerStormstrikes: boolean;
    /**
     * Extra fake players to add. Currently only used by healing sims.
     *
     * @generated from protobuf field: int32 target_dummies = 6
     */
    targetDummies: number;
}
/**
 * @generated from protobuf message proto.SimOptions
 */
export interface SimOptions {
    /**
     * @generated from protobuf field: int32 iterations = 1
     */
    iterations: number;
    /**
     * @generated from protobuf field: int64 random_seed = 2
     */
    randomSeed: bigint;
    /**
     * @generated from protobuf field: bool debug = 3
     */
    debug: boolean; // Enables debug logging.
    /**
     * @generated from protobuf field: bool debug_first_iteration = 6
     */
    debugFirstIteration: boolean;
    /**
     * @generated from protobuf field: bool is_test = 5
     */
    isTest: boolean; // Only used internally.
    /**
     * @generated from protobuf field: bool save_all_values = 7
     */
    saveAllValues: boolean; // Only used internally.
    /**
     * @generated from protobuf field: bool interactive = 8
     */
    interactive: boolean; // Enables interactive mode.
}
// ...
export interface RaidSimRequest {
    /**
     * @generated from protobuf field: proto.Raid raid = 1
     */
    raid?: Raid;
    /**
     * @generated from protobuf field: proto.Encounter encounter = 2
     */
    encounter?: Encounter;
    /**
     * @generated from protobuf field: proto.SimOptions sim_options = 3
     */
    simOptions?: SimOptions;
}
// ...
export const Party = new Party$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Raid$Type extends MessageType<Raid> {
    constructor() {
        super("proto.Raid", [
            { no: 1, name: "parties", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => Party },
            { no: 2, name: "num_active_parties", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 7, name: "buffs", kind: "message", T: () => RaidBuffs },
            { no: 5, name: "debuffs", kind: "message", T: () => Debuffs },
            { no: 4, name: "tanks", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => UnitReference },
            { no: 3, name: "stagger_stormstrikes", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 6, name: "target_dummies", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<Raid>): Raid {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.parties = [];
        message.numActiveParties = 0;
        message.tanks = [];
        message.staggerStormstrikes = false;
        message.targetDummies = 0;
        if (value !== undefined)
            reflectionMergePartial<Raid>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Raid): Raid {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated proto.Party parties */ 1:
                    message.parties.push(Party.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* int32 num_active_parties */ 2:
                    message.numActiveParties = reader.int32();
                    break;
                case /* proto.RaidBuffs buffs */ 7:
                    message.buffs = RaidBuffs.internalBinaryRead(reader, reader.uint32(), options, message.buffs);
                    break;
                case /* proto.Debuffs debuffs */ 5:
                    message.debuffs = Debuffs.internalBinaryRead(reader, reader.uint32(), options, message.debuffs);
                    break;
                case /* repeated proto.UnitReference tanks */ 4:
                    message.tanks.push(UnitReference.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* bool stagger_stormstrikes */ 3:
                    message.staggerStormstrikes = reader.bool();
                    break;
                case /* int32 target_dummies */ 6:
                    message.targetDummies = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Raid, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated proto.Party parties = 1; */
        for (let i = 0; i < message.parties.length; i++)
            Party.internalBinaryWrite(message.parties[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* int32 num_active_parties = 2; */
        if (message.numActiveParties !== 0)
            writer.tag(2, WireType.Varint).int32(message.numActiveParties);
        /* bool stagger_stormstrikes = 3; */
        if (message.staggerStormstrikes !== false)
            writer.tag(3, WireType.Varint).bool(message.staggerStormstrikes);
        /* repeated proto.UnitReference tanks = 4; */
        for (let i = 0; i < message.tanks.length; i++)
            UnitReference.internalBinaryWrite(message.tanks[i], writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        /* proto.Debuffs debuffs = 5; */
        if (message.debuffs)
            Debuffs.internalBinaryWrite(message.debuffs, writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        /* int32 target_dummies = 6; */
        if (message.targetDummies !== 0)
            writer.tag(6, WireType.Varint).int32(message.targetDummies);
        /* proto.RaidBuffs buffs = 7; */
        if (message.buffs)
            RaidBuffs.internalBinaryWrite(message.buffs, writer.tag(7, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message proto.Raid
 */
export const Raid = new Raid$Type();
// ...
export const SimOptions = new SimOptions$Type();
// ...
class RaidSimRequest$Type extends MessageType<RaidSimRequest> {
    constructor() {
        super("proto.RaidSimRequest", [
            { no: 1, name: "raid", kind: "message", T: () => Raid },
            { no: 2, name: "encounter", kind: "message", T: () => Encounter },
            { no: 3, name: "sim_options", kind: "message", T: () => SimOptions }
        ]);
    }
    create(value?: PartialMessage<RaidSimRequest>): RaidSimRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<RaidSimRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RaidSimRequest): RaidSimRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* proto.Raid raid */ 1:
                    message.raid = Raid.internalBinaryRead(reader, reader.uint32(), options, message.raid);
                    break;
                case /* proto.Encounter encounter */ 2:
                    message.encounter = Encounter.internalBinaryRead(reader, reader.uint32(), options, message.encounter);
                    break;
                case /* proto.SimOptions sim_options */ 3:
                    message.simOptions = SimOptions.internalBinaryRead(reader, reader.uint32(), options, message.simOptions);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RaidSimRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* proto.Raid raid = 1; */
        if (message.raid)
            Raid.internalBinaryWrite(message.raid, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* proto.Encounter encounter = 2; */
        if (message.encounter)
            Encounter.internalBinaryWrite(message.encounter, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* proto.SimOptions sim_options = 3; */
        if (message.simOptions)
            SimOptions.internalBinaryWrite(message.simOptions, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message proto.RaidSimRequest
 */
export const RaidSimRequest = new RaidSimRequest$Type();